+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.3.0] - 2022-09-27
 ### Added
 - Warmup-1 delDel
+- Warmup-1 mixStart
 
 ## [1.2.0] - 2022-09-26
 ### Added